ai-apps-collection/agno-apps/react-agno-ai-app/backend/requirements.txt
agno
openai
fastapi
uvicorn[standard]
python-dotenv
tavily-python
yfinance